src/alihub/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
dj-database-url==0.5.0
Django==2.0.8
django-cors-headers==2.4.0
django-countries==5.3.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
Pillow==5.2.0
psycopg2==2.7.6.1
pycparser==2.19
pycryptodomex==3.4.7
PyJWT==1.6.4
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
six==1.11.0
stream-django==1.4.0
stream-python==2.11.0
urllib3==1.23
